Remove PPHHH block again for performance tweaks
We want to optimize HHPPP for now and in the end we can just do the simple copy and mini adjustment that we did before.

Right now, the cost is dominated by the formation of the 2B and 3B bases and generally speaking we are doing a lot of work to find out that the basis size is 0 and we don't need to do anything.

The strategy for the following changes will be to:
1) Bring the contraction out into its own function that we can reuse
2) Prestore HHX and PPP bases and perform cuts as early as possible. We should know that there is no work to do ASAP.
3) TBD